♻️refactor: occurrence 계산 로직 공통화 및 구조 정리
