+++ src/main/java/com/project/backend/domain/briefing/service/query/BriefingQueryServiceImpl.java
@@ -1,18 +1,16 @@
 package com.project.backend.domain.briefing.service.query;
 
 import com.project.backend.domain.briefing.converter.BriefingConverter;
-import com.project.backend.domain.briefing.dto.TodayOccurrenceResult;
+import com.project.backend.domain.occurrence.dto.TodayOccurrenceResult;
 import com.project.backend.domain.briefing.enums.BriefingReason;
-import com.project.backend.domain.event.entity.Event;
 import com.project.backend.domain.event.repository.EventRepository;
 import com.project.backend.domain.briefing.dto.response.BriefingResDTO;
 import com.project.backend.domain.reminder.enums.TargetType;
-import com.project.backend.domain.reminder.provider.OccurrenceProvider;
+import com.project.backend.domain.occurrence.service.OccurrenceResolver;
 import com.project.backend.domain.setting.entity.Setting;
 import com.project.backend.domain.setting.exception.SettingErrorCode;
 import com.project.backend.domain.setting.exception.SettingException;
 import com.project.backend.domain.setting.repository.SettingRepository;
-import com.project.backend.domain.todo.entity.Todo;
 import com.project.backend.domain.todo.repository.TodoRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -35,7 +33,7 @@ public class BriefingQueryServiceImpl implements BriefingQueryService {
     private final SettingRepository settingRepository;
     private final EventRepository eventRepository;
     private final TodoRepository todoRepository;
-    private final OccurrenceProvider occurrenceProvider;
+    private final OccurrenceResolver occurrenceResolver;
 
     public BriefingResDTO.BriefingRes getBriefing(Long memberId) {
         Setting setting = settingRepository.findByMemberId(memberId)
@@ -87,7 +85,7 @@ public class BriefingQueryServiceImpl implements BriefingQueryService {
     private List<BriefingResDTO.BriefInfoRes> toBriefInfos(TargetType type, List<Long> ids, LocalDate date) {
         if (ids.isEmpty()) return List.of();
 
-        return occurrenceProvider.getTodayOccurrence(type, ids, date).stream()
+        return occurrenceResolver.getTodayOccurrence(type, ids, date).stream()
                 .filter(TodayOccurrenceResult::hasToday)
                 .map(BriefingConverter::toBriefInfoRes)
                 .toList();
