+++ src/main/java/com/project/backend/domain/event/service/query/EventQueryService.java
@@ -1,8 +1,8 @@
 package com.project.backend.domain.event.service.query;
 
-import com.project.backend.domain.briefing.dto.TodayOccurrenceResult;
+import com.project.backend.domain.occurrence.dto.TodayOccurrenceResult;
 import com.project.backend.domain.event.dto.response.EventResDTO;
-import com.project.backend.domain.reminder.dto.NextOccurrenceResult;
+import com.project.backend.domain.occurrence.dto.NextOccurrenceResult;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
