+++ src/main/java/com/project/backend/domain/occurrence/service/OccurrenceResolver.java
@@ -1,8 +1,8 @@
-package com.project.backend.domain.reminder.provider;
+package com.project.backend.domain.occurrence.service;
 
-import com.project.backend.domain.briefing.dto.TodayOccurrenceResult;
+import com.project.backend.domain.occurrence.dto.TodayOccurrenceResult;
 import com.project.backend.domain.event.service.query.EventQueryService;
-import com.project.backend.domain.reminder.dto.NextOccurrenceResult;
+import com.project.backend.domain.occurrence.dto.NextOccurrenceResult;
 import com.project.backend.domain.reminder.entity.Reminder;
 import com.project.backend.domain.reminder.enums.TargetType;
 import com.project.backend.domain.todo.service.query.TodoQueryService;
@@ -14,7 +14,7 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class OccurrenceProvider {
+public class OccurrenceResolver {
 
     private final EventQueryService eventQueryService;
     private final TodoQueryService todoService;
